Convert Checkbox components to Typescript
