+++ src/ui/Box.tsx
@@ -95,6 +95,11 @@ type ProgressBarProps = {
   now: number;
 };
 
+type InputProps = {
+  type: string;
+  checked: boolean;
+};
+
 type SpecificComponentProps<T extends TypeaheadModel> = InlineProps &
   Omit<ReactDatePickerProps, 'onChange' | 'onSelect' | 'value' | 'selected'> &
   TitleProps &
@@ -102,6 +107,7 @@ type SpecificComponentProps<T extends TypeaheadModel> = InlineProps &
   LinkProps &
   LabelProps &
   ImageProps &
+  InputProps &
   SvgProps &
   ProgressBarProps &
   Omit<AsyncTypeaheadProps<T>, 'id' | 'size' | 'onChange' | 'selected'>;
@@ -222,7 +228,7 @@ type UIProps = {
   role: UIProp<string>;
 };
 
-type BoxProps<T extends TypeaheadModel = any> = Partial<
+export type BoxProps<T extends TypeaheadModel = any> = Partial<
   UIProps & GeneralProps & SpecificComponentProps<T> & EventHandlerProps
 >;
